refactor: move the PartialOrd impl for unsigned integers
