+++ gadgets/src/traits/utilities/bits/comparator.rs
@@ -14,12 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::utilities::{
-    boolean::Boolean,
-    select::CondSelectGadget,
-    uint::{UInt128, UInt16, UInt32, UInt64, UInt8},
-};
-use snarkvm_fields::{Field, PrimeField};
+use crate::utilities::boolean::Boolean;
+use snarkvm_fields::Field;
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
 
 pub trait EvaluateLtGadget<F: Field> {
@@ -44,50 +40,3 @@ where
         other.less_than_or_equal(cs, self)
     }
 }
-
-macro_rules! uint_cmp_impl {
-    ($($gadget: ident),*) => ($(
-        /*  Bitwise less than comparison of two unsigned integers */
-        impl<F: Field + PrimeField> EvaluateLtGadget<F> for $gadget {
-            fn less_than<CS: ConstraintSystem<F>>(&self, mut cs: CS, other: &Self) -> Result<Boolean, SynthesisError> {
-
-                let mut result = Boolean::constant(true);
-                let mut all_equal = Boolean::constant(true);
-
-                // msb -> lsb
-                for (i, (a, b)) in self
-                    .bits
-                    .iter()
-                    .rev()
-                    .zip(other.bits.iter().rev())
-                    .enumerate()
-                {
-                    // a == 0 & b == 1
-                    let less = Boolean::and(cs.ns(|| format!("not a and b [{}]", i)), &a.not(), b)?;
-
-                    // a == b = !(a ^ b)
-                    let not_equal = Boolean::xor(cs.ns(|| format!("a XOR b [{}]", i)), a, b)?;
-                    let equal = not_equal.not();
-
-                    // evaluate a <= b
-                    let less_or_equal = Boolean::or(cs.ns(|| format!("less or equal [{}]", i)), &less, &equal)?;
-
-                    // select the current result if it is the first bit difference
-                    result = Boolean::conditionally_select(cs.ns(|| format!("select bit [{}]", i)), &all_equal, &less_or_equal, &result)?;
-
-                    // keep track of equal bits
-                    all_equal = Boolean::and(cs.ns(|| format!("accumulate equal [{}]", i)), &all_equal, &equal)?;
-                }
-
-                result = Boolean::and(cs.ns(|| format!("false if all equal")), &result, &all_equal.not())?;
-
-                Ok(result)
-            }
-        }
-
-        /* Bitwise comparison of two unsigned integers */
-        impl<F: Field + PrimeField> ComparatorGadget<F> for $gadget {}
-    )*)
-}
-
-uint_cmp_impl!(UInt8, UInt16, UInt32, UInt64, UInt128);
