+++ gadgets/src/traits/utilities/uint/relational/mod.rs
@@ -14,6 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+pub mod cmp;
+pub use self::cmp::*;
+
 #[macro_use]
 pub mod eq;
 pub use self::eq::*;
