+++ gadgets/src/traits/utilities/uint/uint128.rs
@@ -35,7 +35,7 @@ use snarkvm_utilities::{
     bytes::ToBytes,
 };
 
-use std::{borrow::Borrow, cmp::Ordering};
+use std::borrow::Borrow;
 
 uint_impl_common!(UInt128, u128, 128);
